REFACTOR: More deliberately separate the input binding interface vs implementation since blas fucked it all up. Functions and structs were moved from the interface into the main crate.
