+++ crates/thetawave_interface/Cargo.toml
@@ -15,7 +15,5 @@ bevy_ecs = { workspace = true }
 bevy_reflect = { workspace = true }
 leafwing-input-manager = { workspace = true }
 bevy_input = { workspace = true }
-ron = { workspace = true }
 bevy_time = { workspace = true }
 bevy_math = { workspace = true }
-rand = { workspace = true }
