+++ crates/thetawave_interface/src/lib.rs
@@ -3,8 +3,8 @@ pub mod character;
 pub mod character_selection;
 pub mod game;
 pub mod health;
+pub mod input;
 pub mod objective;
-pub mod options;
 pub mod player;
 pub mod run;
 pub mod spawnable;
