+++ src/options/mod.rs
@@ -1,13 +1,17 @@
 //! `thetawave` player module
 use bevy::prelude::*;
 use leafwing_input_manager::prelude::InputManagerPlugin;
-use thetawave_interface::options::input::{get_input_bindings, InputsResource, MenuAction};
+use thetawave_interface::input::{InputsResource, MenuAction};
 
 mod display;
 mod input;
 
 use crate::states;
+use input::get_input_bindings;
 use std::default::Default;
+use std::env::current_dir;
+use std::fs::{DirBuilder, File};
+use std::io::Write;
 use std::path::PathBuf;
 
 pub use self::display::{
@@ -44,7 +48,7 @@ impl Plugin for OptionsPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins(InputManagerPlugin::<MenuAction>::default());
 
-        app.insert_resource(InputsResource::new(get_input_bindings()));
+        app.insert_resource(InputsResource::from(get_input_bindings()));
 
         #[cfg(not(target_arch = "wasm32"))]
         app.add_systems(Startup, (set_window_icon, spawn_menu_explorer_system));
@@ -54,6 +58,51 @@ impl Plugin for OptionsPlugin {
         app.add_systems(
             Update,
             toggle_zoom_system.run_if(in_state(states::AppStates::Game)),
+        );
+    }
+}
+
+/// Creates config file in config directory from config file in this directory
+macro_rules! confgen {
+    ( $($filename:expr),* ) => {
+        {
+            let conf_dir = current_dir().unwrap().join("config");
+            if !conf_dir.is_dir() {
+                DirBuilder::new()
+                    .create(conf_dir.clone())
+                    .expect("Confgen failed: could not create config dir.");
+            }
+
+            $({
+                let default = include_bytes!($filename);
+                let file_path = conf_dir.join($filename);
+                if !file_path.is_file() {
+                    let mut file = File::create(file_path)
+                        .expect(concat!("Confgen failed: could not create config file ", $filename, "."));
+                    file.write_all(default)
+                        .expect(concat!("Confgen failed: could not write config file ", $filename, "."));
+                }
+            })*
+        }
+    }
+}
+
+/// Generates the display config file
+pub fn generate_config_files() {
+    confgen!("display.ron");
+    confgen!("input.ron");
+}
+
+#[cfg(all(test, not(target_arch = "wasm32"), feature = "cli"))]
+mod cli_tests {
+    use argh::FromArgs;
+    #[test]
+    fn test_cli_parse_asset_path_dir() {
+        assert_eq!(
+            super::GameInitCLIOptions::from_args(&["thetawave"], &["--assets-dir", "myassets/"])
+                .unwrap()
+                .assets_dir,
+            Some(std::path::PathBuf::from("myassets/"))
         );
     }
 }
