+++ src/player/systems/ability.rs
@@ -1,9 +1,9 @@
 use bevy::prelude::*;
 use bevy_rapier2d::prelude::{ExternalImpulse, Velocity};
 use leafwing_input_manager::prelude::ActionState;
+use thetawave_interface::input::PlayerAction;
 use thetawave_interface::{
     audio::{PlaySoundEffectEvent, SoundEffectType},
-    options::input::PlayerAction,
     player::{AbilityType, PlayerComponent},
 };
